# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para cadastrar domínios/urls e verificar o status http de cada um. Cada verificação também é gravada no arquivo de resultados.",
        "title": "Uptime Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Retorna ok quando a api está no ar",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Healthcheck",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/urls": {
            "get": {
                "description": "Faz um GET em todas as urls cadastradas, retorna o status http de cada uma e grava os resultados no arquivo de resultados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Verifica todas as urls",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CheckResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cadastra uma url para ser monitorada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Cadastra uma url",
                "parameters": [
                    {
                        "description": "Dados da url",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateURLRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.URL"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/urls/{id}": {
            "get": {
                "description": "Faz um GET apenas na url do id informado, retorna o status http e grava o resultado no arquivo de resultados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Verifica uma url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da url (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CheckResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CheckResponse": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "type": "string",
                    "example": "2026-07-20T01:17:58Z"
                },
                "duration_ms": {
                    "type": "integer",
                    "example": 478
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": "84c142da-dd0f-4664-bd43-0e32ff1c5ef7"
                },
                "name": {
                    "type": "string",
                    "example": "google"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "url": {
                    "type": "string",
                    "example": "https://www.google.com"
                }
            }
        },
        "models.CreateURLRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "google"
                },
                "url": {
                    "type": "string",
                    "example": "https://www.google.com"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "url não encontrada"
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.URL": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-20T01:16:23Z"
                },
                "id": {
                    "type": "string",
                    "example": "84c142da-dd0f-4664-bd43-0e32ff1c5ef7"
                },
                "name": {
                    "type": "string",
                    "example": "google"
                },
                "url": {
                    "type": "string",
                    "example": "https://www.google.com"
                }
            }
        }
    }
}